Code review suggested changes.
